Use unsafe pointers in raw allocator to fix potential alignment problems (#19)
* Use unsafe pointers in raw allocator to fix potential alignment problems

* Fix clear bug and generated code

* Update linter to 1.30.0 version

* Fix performance degradation in Clear

+++ generator/internal/testdata/expected/circlecolor.alloc.go
@@ -75,13 +75,13 @@ func (s CircleColorBuffer) SubSlice(low int, high int) CircleColorBuffer {
 	var tVar CircleColor
 	tSize := unsafe.Sizeof(tVar)
 	type internalPtr struct {
-		offset    uint32
+		offset    uintptr
 		bucketIdx uint8
 		arenaMask uint16
 	}
 	currentPtr := *(*internalPtr)(unsafe.Pointer(&s.data))
 	newPtr := internalPtr{
-		offset:    currentPtr.offset + uint32(low*int(tSize)),
+		offset:    currentPtr.offset + uintptr(low*int(tSize)),
 		bucketIdx: currentPtr.bucketIdx,
 		arenaMask: currentPtr.arenaMask,
 	}
@@ -105,13 +105,13 @@ func (s CircleColorBuffer) Get(idx int) CircleColorPtr {
 	var tVar CircleColor
 	tSize := unsafe.Sizeof(tVar)
 	type internalPtr struct {
-		offset    uint32
+		offset    uintptr
 		bucketIdx uint8
 		arenaMask uint16
 	}
 	currentPtr := *(*internalPtr)(unsafe.Pointer(&s.data))
 	newPtr := internalPtr{
-		offset:    currentPtr.offset + uint32(idx*int(tSize)),
+		offset:    currentPtr.offset + uintptr(idx*int(tSize)),
 		bucketIdx: currentPtr.bucketIdx,
 		arenaMask: currentPtr.arenaMask,
 	}
